Move the update-support-docs implementation into xtask
The docs/support.md rendering lived in dewasm-cli behind a [lib]
target that existed only for `cargo xtask update-support-docs` and
the compare-only freshness test. The CLI crate is the dewasm binary;
a docs-regeneration task has no reason to live there, and it forced
xtask to depend on dewasm-cli. Move the rendering next to the
command, keep the freshness check as a unit test beside it, and drop
the CLI crate's [lib] target and xtask's dewasm-cli dependency.
